LayoutTests/http/tests/media/video-play-stall-seek-expected.txt

Test that playback can be resumed by seeking backwards after load stalls.

EVENT(waiting)
TEST(video.readyState == HTMLMediaElement.DATA_UNAVAILABLE) OK
RUN(video.currentTime = 0.5)
EVENT(canplay)
TEST(video.readyState == HTMLMediaElement.CAN_PLAY) OK
TEST(video.networkState == HTMLMediaElement.LOADED_FIRST_FRAME) OK
END OF TEST
